requirements.txt
facenet-pytorch
pillow
ultralytics
moviepy
keyboard